PVP-250 Rename Analysis to Statistics
