+++ app/src/main/java/com/pvp/app/common/DateUtil.kt
@@ -104,7 +104,8 @@ object DateUtil {
 
     fun Timestamp.toLocalDate(): LocalDate {
         return LocalDate.ofInstant(
-            toDate().toInstant(),
+            toDate()
+                .toInstant(),
             ZoneId.systemDefault()
         )
     }
