+++ app/src/main/res/values/strings.xml
@@ -9,7 +9,6 @@
     <string name="form_create_meal_preparation">Preparation</string>
     <string name="form_create_meal_title">Meal title</string>
     <string name="form_survey_toast_error">Submission was not successful</string>
-    <string name="route_analysis">Analysis</string>
     <string name="route_calendar">Calendar</string>
     <string name="route_decorations">Decorations</string>
     <string name="route_friend">%1$s Profile</string>
@@ -17,6 +16,7 @@
     <string name="route_goals">Goals</string>
     <string name="route_profile">Profile</string>
     <string name="route_settings">Settings</string>
+    <string name="route_statistics">Statistics</string>
     <string name="route_steps">Steps</string>
     <string name="route_survey">Survey</string>
     <string name="screen_profile_button_sign_out">Sign Out</string>
